app/requirements.txt
asgiref==3.5.2
Django==4.1.4
sqlparse==0.4.3
tzdata==2022.7
psycopg2-binary==2.9.4
djangorestframework==3.13.1
pytest==6.2.5
pytest-django==4.4.0
pytest-factoryboy
Pillow
gunicorn==20.1.0